+++ packages/core/src/ExecuteScriptRoutine.ts
@@ -4,11 +4,11 @@
  */
 
 import path from 'path';
-import { ModuleLoader, Routine } from '@boost/core';
+import { Routine, Task } from '@boost/core';
 import parseArgs from 'yargs-parser';
 import Script from './Script';
 import ScriptContext from './contexts/ScriptContext';
-import { BeemoTool, Execution } from './types';
+import { BeemoTool, ExecuteType } from './types';
 
 export default class ExecuteScriptRoutine extends Routine<ScriptContext, BeemoTool> {
   bootstrap() {
@@ -16,29 +16,51 @@ export default class ExecuteScriptRoutine extends Routine<ScriptContext, BeemoTo
     this.task(this.tool.msg('app:scriptRun'), this.runScript);
   }
 
-  execute(): Promise<Execution> {
+  execute(): Promise<any> {
     return this.serializeTasks();
   }
 
   /**
    * Attempt to load a script from the configuration module.
    */
-  async loadScript(context: ScriptContext): Promise<Script> {
-    const filePath = path.join(context.moduleRoot, 'scripts', `${context.scriptName}.js`);
-    const loader = new ModuleLoader(this.tool, 'script', Script);
+  loadScript(context: ScriptContext): Script {
+    const { loader } = this.tool.getRegisteredPlugin('script');
+    let script: Script;
 
-    this.debug('Loading script');
+    // Try file path in configuration module
+    try {
+      this.debug('Loading script from configuration module');
 
-    const script = loader.importModule(filePath, [
-      context.scriptName,
-      this.tool.msg('app:scriptRunNamed', { name: context.scriptName }),
-    ]);
+      const filePath = path.join(context.moduleRoot, 'scripts', `${context.scriptName}.js`);
 
-    // Pass context and tool to script
-    script.configure(this);
+      script = loader.importModule(filePath);
+      script.name = context.scriptName;
 
-    // Set script into context
-    context.setScript(script, filePath);
+      context.setScript(script, filePath);
+    } catch (error1) {
+      // Try an NPM module
+      try {
+        this.debug('Loading script from NPM module');
+
+        script = loader.importModule(context.eventName); // Module names are kebab case
+
+        context.setScript(
+          script,
+          // Cannot mock require.resolve in Jest
+          process.env.NODE_ENV === 'test' ? script.moduleName : require.resolve(script.moduleName),
+        );
+      } catch (error2) {
+        throw new Error(
+          [
+            'Failed to load script, the following errors occurred:',
+            error1.message,
+            error2.message,
+          ].join('\n'),
+        );
+      }
+    }
+
+    this.tool.addPlugin('script', script);
 
     this.tool.emit(`${context.eventName}.load-script`, [context, script]);
 
@@ -48,7 +70,7 @@ export default class ExecuteScriptRoutine extends Routine<ScriptContext, BeemoTo
   /**
    * Run the script while also parsing arguments to use as options.
    */
-  async runScript(context: ScriptContext, script: Script): Promise<Execution> {
+  async runScript(context: ScriptContext, script: Script): Promise<any> {
     const { argv } = this.context;
 
     this.debug('Executing script with args "%s"', argv.join(' '));
@@ -61,6 +83,10 @@ export default class ExecuteScriptRoutine extends Routine<ScriptContext, BeemoTo
     try {
       result = await script.execute(context, args);
 
+      if (typeof result === 'object' && result && result.type && Array.isArray(result.tasks)) {
+        result = await this.runScriptTasks(args, result.type, result.tasks);
+      }
+
       this.tool.emit(`${context.eventName}.after-execute`, [context, result, script]);
     } catch (error) {
       this.tool.emit(`${context.eventName}.failed-execute`, [context, error, script]);
@@ -69,5 +95,26 @@ export default class ExecuteScriptRoutine extends Routine<ScriptContext, BeemoTo
     }
 
     return result;
+  }
+
+  /**
+   * Run the tasks the script enqueued using the defined process.
+   */
+  async runScriptTasks(args: any, type: ExecuteType, tasks: Task<any>[]): Promise<any> {
+    // Add the tasks to the routine so they show in the console
+    this.tasks.push(...tasks);
+
+    switch (type) {
+      case 'parallel':
+        return this.parallelizeTasks(args, tasks);
+      case 'pool':
+        return this.poolTasks(args, {}, tasks);
+      case 'serial':
+        return this.serializeTasks(args, tasks);
+      case 'sync':
+        return this.synchronizeTasks(args, tasks);
+      default:
+        throw new Error(`Unknown execution type "${type}"`);
+    }
   }
 }