+++ packages/core/src/Script.ts
@@ -3,12 +3,16 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
-import { Routine } from '@boost/core';
+import { Plugin, Task } from '@boost/core';
 import { Options } from 'yargs-parser';
 import ScriptContext from './contexts/ScriptContext';
-import { BeemoTool } from './types';
+import { ExecuteType } from './types';
 
-export default class Script extends Routine<ScriptContext, BeemoTool> {
+export default class Script<Args extends object = {}, Opts extends object = {}> extends Plugin<
+  Opts
+> {
+  tasks: Task<any>[] = [];
+
   /**
    * Define a configuration object to parse args with.
    */
@@ -17,10 +21,34 @@ export default class Script extends Routine<ScriptContext, BeemoTool> {
   }
 
   /**
-   * No-op.
+   * Execute the script with the context and parsed args.
    */
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  execute(context: ScriptContext, value: any): Promise<any> {
-    return Promise.resolve();
+  execute(context: ScriptContext, args: Args): Promise<any> {
+    return this.executeTasks('serial');
+  }
+
+  /**
+   * Execute the enqueued tasks using the defined process.
+   */
+  executeTasks(type: ExecuteType) {
+    return Promise.resolve({
+      tasks: this.tasks,
+      type,
+    });
+  }
+
+  /**
+   * Define an individual task that will be piped to an upstream routine.
+   */
+  task(title: string, action: any): Task<any> {
+    if (typeof action !== 'function') {
+      throw new TypeError(this.tool.msg('errors:taskRequireAction'));
+    }
+
+    const task = new Task(title, action);
+
+    this.tasks.push(task);
+
+    return task;
   }
 }