+++ packages/core/tests/Script.test.ts
@@ -1,21 +1,61 @@
+import { mockTool } from '@boost/test-utils';
 import Script from '../src/Script';
+import { createScriptContext } from '../../../tests/helpers';
 
 describe('Script', () => {
   let script: Script;
 
-  class TestScript extends Script {
-    execute() {
-      return Promise.resolve();
-    }
-  }
-
   beforeEach(() => {
-    script = new TestScript('test', 'test');
+    script = new Script();
+    script.tool = mockTool();
   });
 
   describe('args()', () => {
     it('returns an empty object', () => {
       expect(script.args()).toEqual({});
+    });
+  });
+
+  describe('execute()', () => {
+    it('executes tasks serially by default', async () => {
+      const spy = jest.spyOn(script, 'executeTasks');
+
+      await script.execute(createScriptContext(), {});
+
+      expect(spy).toHaveBeenCalledWith('serial');
+    });
+  });
+
+  describe('executeTasks()', () => {
+    it('returns tasks and process type', async () => {
+      script.task('foo', () => 123);
+
+      const result = await script.executeTasks('parallel');
+
+      expect(result).toEqual({
+        tasks: script.tasks,
+        type: 'parallel',
+      });
+    });
+  });
+
+  describe('task()', () => {
+    it('errors if not a function', () => {
+      expect(() => {
+        // @ts-ignore Allow non-function
+        script.task('foo', 'bar');
+      }).toThrowErrorMatchingSnapshot();
+    });
+
+    it('maps `Task` objects', () => {
+      expect(script.tasks).toHaveLength(0);
+
+      script.task('foo', () => 123);
+      script.task('bar', () => 123);
+
+      expect(script.tasks).toHaveLength(2);
+      expect(script.tasks[0].constructor.name).toBe('Task');
+      expect(script.tasks[1].constructor.name).toBe('Task');
     });
   });
 });